refactor(agent): consolidate agent setup into BrowserOS AI settings
- Remove the "Recent agents" / "Manage agents" section from the home page.
  Agents stay reachable via the composer's provider picker. Drops the now
  dead AgentCardDock, HomeAgentCard, and orderHomeAgents helpers.
- Remove the "New" badges on the ChatGPT/Copilot/Qwen provider template
  cards and the "New" pill on the MCP promo banner.
- Bring Claude Code + Codex into the BrowserOS AI pane as a "Coding agents"
  section: each card opens the New Agent dialog scoped to that runtime, and
  existing agents are managed inline. Removes the separate Claude Code /
  Codex settings tabs (AISettingsPage now renders the single pane) and the
  per-adapter AdapterAgentsPane / section-routing machinery.
- Hide the redundant Adapter picker in NewAgentDialog when scoped to a
  single adapter; point the legacy /agents redirect at /settings/ai.
