+++ packages/browseros-agent/apps/agent/entrypoints/app/ai-settings/ProviderTemplatesSection.tsx
@@ -55,20 +55,13 @@ export const ProviderTemplatesSection: FC<ProviderTemplatesSectionProps> = ({
 
         <CollapsibleContent>
           <div className="mt-4 grid gap-3 sm:grid-cols-2 lg:grid-cols-3">
-            {filteredTemplates.map((template) => {
-              const isNew =
-                template.id === 'chatgpt-pro' ||
-                template.id === 'github-copilot' ||
-                template.id === 'qwen-code'
-              return (
-                <ProviderTemplateCard
-                  key={template.id}
-                  template={template}
-                  isNew={isNew}
-                  onUseTemplate={onUseTemplate}
-                />
-              )
-            })}
+            {filteredTemplates.map((template) => (
+              <ProviderTemplateCard
+                key={template.id}
+                template={template}
+                onUseTemplate={onUseTemplate}
+              />
+            ))}
           </div>
         </CollapsibleContent>
       </div>
